ヘッダファイルやソースファイルのフィルタ分けを行った。 警告が出ないようにソースコードやプリプロセッサ、リンカーの編集などをした。

+++ 3D就活ゲーム制作/GameProgramming/CMatrix.cpp
@@ -39,7 +39,7 @@ CMatrix CMatrix::Identity() {
 //RotateY(角度)
 CMatrix CMatrix::RotateY(float degree) {
 	//角度からラジアンを求める
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	//単位行列にする
 	Identity();
 	//Y軸で回転する行列の設定
@@ -51,7 +51,7 @@ CMatrix CMatrix::RotateY(float degree) {
 }
 
 CMatrix CMatrix::RotateX(float degree) {
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	Identity();
 	mM[1][1] = mM[2][2] = cosf(rad);
 	mM[1][2] = sinf(rad);
@@ -60,7 +60,7 @@ CMatrix CMatrix::RotateX(float degree) {
 }
 
 CMatrix CMatrix::RotateZ(float degree) {
-	float rad = degree / 180.0f * M_PI;
+	float rad = degree / 180.0f * (float)M_PI;
 	Identity();
 	mM[0][0] = mM[1][1] = cosf(rad);
 	mM[0][1] = sinf(rad);
